semantic_evaluation/requirements.txt
ipython==9.6.0
pyenchant==3.3.0
